# swagger.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081/api","description":"Generated server url"}],"paths":{"/associate/{id}":{"get":{"tags":["associate-controller"],"operationId":"findById_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AssociateDetailedResponse"}}}}}},"put":{"tags":["associate-controller"],"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssociateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AssociateDetailedResponse"}}}}}}},"/agenda/{id}":{"get":{"tags":["agenda-controller"],"operationId":"findById_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AgendaDetailedResponse"}}}}}},"put":{"tags":["agenda-controller"],"operationId":"update_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgendaRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AgendaDetailedResponse"}}}}}},"delete":{"tags":["agenda-controller"],"operationId":"disable","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AgendaDetailedResponse"}}}}}}},"/voting-session":{"get":{"tags":["voting-session-controller"],"operationId":"findPage","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseVotingSessionResponse"}}}}}},"post":{"tags":["voting-session-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VotingSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/VotingSessionResponse"}}}}}}},"/voting-session/{id}/vote":{"post":{"tags":["voting-session-controller"],"operationId":"createVote","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VoteRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/VoteResponse"}}}}}}},"/associate":{"get":{"tags":["associate-controller"],"operationId":"findPage_1","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseAssociateResponse"}}}}}},"post":{"tags":["associate-controller"],"operationId":"create_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AssociateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AssociateDetailedResponse"}}}}}}},"/agenda":{"get":{"tags":["agenda-controller"],"operationId":"findPage_2","parameters":[{"name":"pageable","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/PaginatedResponseAgendaResponse"}}}}}},"post":{"tags":["agenda-controller"],"operationId":"create_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AgendaRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/AgendaDetailedResponse"}}}}}}},"/voting-session/{id}":{"get":{"tags":["voting-session-controller"],"operationId":"findById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/VotingSessionResponse"}}}}}},"patch":{"tags":["voting-session-controller"],"operationId":"update_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VotingSessionPatchRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/vnd.assembly.api.v1+json":{"schema":{"$ref":"#/components/schemas/VotingSessionResponse"}}}}}}}},"components":{"schemas":{"AssociateRequest":{"required":["document","name"],"type":"object","properties":{"name":{"maxLength":30,"minLength":0,"type":"string"},"document":{"maxLength":11,"minLength":11,"pattern":"^([0-9]{11})$","type":"string"}}},"AgendaResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"AssociateDetailedResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"document":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"agendas":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AgendaResponse"}}}},"AgendaRequest":{"required":["description","name"],"type":"object","properties":{"name":{"maxLength":30,"minLength":0,"type":"string"},"description":{"type":"string"},"associateIds":{"uniqueItems":true,"type":"array","items":{"type":"integer","format":"int64"}}}},"AgendaDetailedResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"},"associates":{"uniqueItems":true,"type":"array","items":{"$ref":"#/components/schemas/AssociateResponse"}}}},"AssociateResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"VotingSessionRequest":{"required":["agendaId","startsAt"],"type":"object","properties":{"agendaId":{"type":"integer","format":"int64"},"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"}}},"VotingSessionResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"agendaId":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"},"totalVotes":{"type":"integer","format":"int64"},"yesVotes":{"type":"integer","format":"int64"},"noVotes":{"type":"integer","format":"int64"},"result":{"type":"string"}}},"VoteRequest":{"required":["associateId","content"],"type":"object","properties":{"associateId":{"type":"integer","format":"int64"},"content":{"type":"boolean"}}},"VoteResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"associateId":{"type":"integer","format":"int64"},"content":{"type":"boolean"},"votingSessionId":{"type":"integer","format":"int64"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"VotingSessionPatchRequest":{"required":["startsAt"],"type":"object","properties":{"startsAt":{"type":"string","format":"date-time"},"endsAt":{"type":"string","format":"date-time"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PaginatedResponseVotingSessionResponse":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/VotingSessionResponse"}}}},"PaginatedResponseAssociateResponse":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AssociateResponse"}}}},"PaginatedResponseAgendaResponse":{"type":"object","properties":{"number":{"type":"integer","format":"int32"},"size":{"type":"integer","format":"int32"},"totalPages":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/AgendaResponse"}}}}}}}